use lldb from installer (#8047)
* use lldb from installer

since 2025.3 CLion prefers to use lldbfrontend from
native-debug-plugin and soon lldbfrontend will be removed
from the CLion installer. Unfortunately, our tests are run
with the flat classpath and we can't rely on built-in code
do detect lldbfrontend based on some NDP classloader.

let's rely on the flag to use installer-provided lldbfrontend
and decide on the next steps later one it's deprecated

* Disable soft_fail for clion-oss-under-dev

All the issues are addressed and we can have soft_fail as false now