Add RadBazelProjectFallbackConfigDisabler (#8055)
We found that providing fallback configuration to the backend leads to recursive watch roots being set up for include paths mentioned in the fallback config. We know this has very negative performance implications because of greediness in VfsWarmQueue. This extension point allows us to disable fallback configurations for Bazel projects.

However, since the extension point was introduced in 2025.2.5, it is not save to simply instantiate the implementation class and we need to do some manual registration magic.